refactor: rename welcomewrapper
