+++ packages/dapp/src/App.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { AppProvider } from './AppProvider';
-import { FeatureCards, WelcomeWrapper } from './components';
+import { FeatureCards, WelcomeScreen } from './components';
 import { Navbar } from './components/Navbar';
 import { Menus } from './components/Navbar/Menu/Menus';
 import { Widgets } from './components/Widgets';
@@ -19,7 +19,7 @@ export default function App() {
     <AppProvider>
       <Navbar />
       <Menus />
-      <WelcomeWrapper />
+      <WelcomeScreen />
       <Widgets />
       <FeatureCards />
     </AppProvider>
