+++ packages/dapp/src/components/WelcomeScreen/index.ts
@@ -1,0 +1,1 @@
+export * from './WelcomeScreen';
