--- packages/dapp/src/components/WelcomeWrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './WelcomeWrapper';
